+++ packages/api/src/business/processes/__mocks__/titres-activites-statut-ids-update-activites.ts
@@ -5,7 +5,7 @@ const titresActivitesDelaiDepasse = [
   {
     date: '1000-01-01',
     activiteStatutId: ACTIVITES_STATUTS_IDS.ABSENT,
-    type: { delaiMois: 3 },
+    typeId: 'gra',
   },
 ] as TitresActivites[]
 
@@ -13,7 +13,7 @@ const titresActivitesDelaiNonDepasse = [
   {
     date: '3000-01-01',
     activiteStatutId: ACTIVITES_STATUTS_IDS.ABSENT,
-    type: { delaiMois: 3 },
+    typeId: 'gra',
   },
 ] as TitresActivites[]
