--- packages/api/src/business/processes/__mocks__/titres-activites-update-titres.ts
@@ -1,34 +1,0 @@
-import { newTitreId } from '../../../database/models/_format/id-create.js'
-import ActivitesTypes from '../../../database/models/activites-types.js'
-import Titres from '../../../database/models/titres.js'
-
-const titresSansActivite = [
-  {
-    id: newTitreId('h-cx-courdemanges-1988'),
-    activites: [],
-    titulaires: [{ utilisateurs: [{ email: 'email' }] }],
-  },
-] as unknown as Titres[]
-
-const titresToutesActivites = [
-  {
-    id: newTitreId('h-cx-courdemanges-1988'),
-    activites: [
-      { annee: 2018, periodeId: 1 },
-      { annee: 2018, periodeId: 2 },
-      { annee: 2018, periodeId: 3 },
-      { annee: 2018, periodeId: 4 },
-    ],
-  },
-] as Titres[]
-
-const titreActiviteTypeGuyane = {
-  id: 'grp',
-  dateDebut: 2018,
-  nom: "rapport d'activité",
-  frequenceId: 'tri',
-} as unknown as ActivitesTypes
-
-const titreActivitesTypes = [titreActiviteTypeGuyane]
-
-export { titresSansActivite, titresToutesActivites, titreActivitesTypes }
