+++ packages/api/src/business/rules/__mocks__/titre-activite-statut-id-find-activites.ts
@@ -4,23 +4,23 @@ import { ACTIVITES_STATUTS_IDS } from 'camino-common/src/static/activitesStatuts
 export const titreActiviteFermee = {
   activiteStatutId: ACTIVITES_STATUTS_IDS.CLOTURE,
   date: '1000-01-01',
-  type: { delaiMois: 3 },
+  typeId: 'gra',
 } as ITitreActivite
 
 export const titreActiviteDeposee = {
   activiteStatutId: ACTIVITES_STATUTS_IDS.DEPOSE,
   date: '1000-01-01',
-  type: { delaiMois: 3 },
+  typeId: 'gra',
 } as ITitreActivite
 
 export const titreActiviteAbsenteDelaiDepasse = {
   activiteStatutId: ACTIVITES_STATUTS_IDS.ABSENT,
   date: '1000-01-01',
-  type: { delaiMois: 3 },
+  typeId: 'gra',
 } as ITitreActivite
 
 export const titreActiviteEnCoursDelaiNonDepasse = {
   activiteStatutId: ACTIVITES_STATUTS_IDS.EN_CONSTRUCTION,
   date: '3000-01-01',
-  type: { delaiMois: 3 },
+  typeId: 'gra',
 } as ITitreActivite
